Update functions.md - New Examples and shorter intro
The fucntions were going too much into "what it is" instead of "what they are for" or "what can be done with them".

I added a new intro paragraph, but kept the previous one. I think they are working well together but feel free to make changes.

I also added a new section [Examples] and wrote some example use-cases to better explain what is possible with the tool at hand.

You might have better ideas for use cases to list here, feel free to change anything. My intention here was less to write a perfect text and more to suggest a different approach to help users better understand how powerful this feature is and a bit of what is possible. Especially since this is web3 and some users might be used to the limitations on other chains.